requirements.txt
python-telegram-bot
autopy